added map controls
