+++ frontend/package.json
@@ -4,6 +4,7 @@
   "description": "",
   "private": true,
   "dependencies": {
+    "@mapbox/mapbox-gl-draw": "^1.3.0",
     "animate.css": "^4.1.1",
     "bootstrap": "^4.6.0",
     "classnames": "^2.3.1",
